ENH: Updating MANIFEST to include all data formats as well as moving test into test_io
