+++ MANIFEST.in
@@ -10,7 +10,7 @@ recursive-exclude * *.py[co]
 recursive-include act/plotting *.txt
 
 recursive-include docs *.rst conf.py Makefile make.bat
-recursive-include act/tests/data *.cdf *.nc *.data
+recursive-include act/tests/data *
 
 include versioneer.py
 include act/_version.py
