+++ act/tests/sample_files.py
@@ -33,6 +33,10 @@ be used for testing ACT.
     EXAMPLE_CO2FLX4M
     EXAMPLE_SIRS
     EXAMPLE_GML_RADIATION
+    EXAMPLE_GML_MET
+    EXAMPLE_GML_OZONE
+    EXAMPLE_GML_CO2
+    EXAMPLE_GML_HALO
 """
 import os
